// cgocore-ui/src/app/features/models/reference/airlines/views/full/AirlineCompanyFullView.ts
import { AirlineCompanyForm } from 'src/app/features/models/reference/airlines/forms/AirlineCompanyForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirlineCompanyFullView {

id: number;
form: AirlineCompanyForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}